Issue #OS-81 Move APImessage to Pojo
