+++ java/pojos/src/main/java/io/opensaber/pojos/APIMessage.java
@@ -1,7 +1,6 @@
-package io.opensaber.registry.interceptor.handler;
+package io.opensaber.pojos;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.opensaber.pojos.Request;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
