+++ java/pojos/src/main/java/io/opensaber/pojos/RequestWrapper.java
@@ -1,11 +1,7 @@
-package io.opensaber.registry.interceptor.handler;
+package io.opensaber.pojos;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.annotation.RequestScope;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
