+++ java/registry-interceptor/src/main/java/io/opensaber/registry/interceptor/handler/BaseResponseHandler.java
@@ -8,17 +8,13 @@ import java.util.Map;
 
 import javax.servlet.http.HttpServletResponse;
 
+import io.opensaber.pojos.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
-
-import io.opensaber.pojos.Response;
-import io.opensaber.pojos.ResponseParams;
-import io.opensaber.pojos.ValidationResponse;
-import io.opensaber.pojos.ValidationResponseSerializer;
 
 /**
  *
